+++ docs/src/components/ReactExample/components/Playground.tsx
@@ -39,7 +39,7 @@ const Playground = ({ exampleKnobs, onChange }: Props) => {
     return defaultVals;
   });
 
-  const changeKnob = ({
+  const handleChangeKnob = ({
     component,
     name,
     value,
@@ -71,20 +71,24 @@ const Playground = ({ exampleKnobs, onChange }: Props) => {
             </Text>
             <StyledKnobsWrapper>
               {sortBy(knobs, ["type"]).map(({ type, name, options }) => {
-                const fieldName = `${component}-${name}`;
+                const baseKnobProps = {
+                  name: `${component}-${name}`,
+                  value: values[component][name],
+                };
+
                 if (type === "boolean") {
                   return (
                     <BooleanKnob
                       key={name}
                       checked={values[component][name]}
                       onChange={ev =>
-                        changeKnob({
+                        handleChangeKnob({
                           component,
                           name,
                           value: ev.target.checked,
                         })
                       }
-                      name={fieldName}
+                      {...baseKnobProps}
                     />
                   );
                 }
@@ -93,16 +97,15 @@ const Playground = ({ exampleKnobs, onChange }: Props) => {
                   return (
                     <SelectKnob
                       key={name}
-                      value={values[component][name]}
                       onChange={ev =>
-                        changeKnob({
+                        handleChangeKnob({
                           component,
                           name,
                           value: ev.target.value,
                         })
                       }
-                      name={fieldName}
                       options={options}
+                      {...baseKnobProps}
                     />
                   );
                 }
@@ -111,17 +114,16 @@ const Playground = ({ exampleKnobs, onChange }: Props) => {
                   return (
                     <IconKnob
                       key={name}
-                      value={values[component][name]}
-                      name={fieldName}
                       onChange={ev => {
                         const { value } = ev.target;
                         const val = value !== "null" ? `${value}-icon` : null;
-                        changeKnob({
+                        handleChangeKnob({
                           component,
                           name,
                           value: val,
                         });
                       }}
+                      {...baseKnobProps}
                     />
                   );
                 }
@@ -130,15 +132,14 @@ const Playground = ({ exampleKnobs, onChange }: Props) => {
                   return (
                     <NumberKnob
                       key={name}
-                      value={values[component][name]}
-                      name={fieldName}
                       onChange={ev =>
-                        changeKnob({
+                        handleChangeKnob({
                           component,
                           name,
                           value: parseInt(ev.target.value, 10),
                         })
                       }
+                      {...baseKnobProps}
                     />
                   );
                 }
@@ -146,15 +147,14 @@ const Playground = ({ exampleKnobs, onChange }: Props) => {
                 return (
                   <TextKnob
                     key={name}
-                    value={values[component][name]}
-                    onChange={ev =>
-                      changeKnob({
+                    {...baseKnobProps}
+                    onChange={ev => {
+                      handleChangeKnob({
                         component,
                         name,
                         value: ev.target.value,
-                      })
-                    }
-                    name={fieldName}
+                      });
+                    }}
                   />
                 );
               })}